Revert "refactor: migrate testUtils to client"
This reverts commit ae569e2021a5c720111518cf3f2236e761a7186d.
