+++ packages/core/client/src/schema-component/antd/action/__tests__/action.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { fireEvent, render, screen, sleep, userEvent } from '../../../../test';
+import { fireEvent, render, screen, sleep, userEvent } from 'testUtils';
 import App1 from '../demos/demo1';
 import App2 from '../demos/demo2';
 import App3 from '../demos/demo3';
