+++ packages/core/client/src/schema-component/antd/icon-picker/__tests__/icon-picker.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render, screen, userEvent } from '../../../../test';
+import { render, screen, userEvent } from 'testUtils';
 import App from '../demos/icon-picker';
 
 describe('IconPicker', () => {
