+++ packages/core/client/src/schema-component/antd/input-number/__tests__/input-number.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { fireEvent, render, screen } from '../../../../test';
+import { fireEvent, render, screen } from 'testUtils';
 import App2 from '../demos/addonBefore&addonAfter';
 import App3 from '../demos/highPrecisionDecimals';
 import App1 from '../demos/inputNumber';
