+++ packages/core/client/src/schema-component/antd/markdown/__tests__/markdown.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { act, fireEvent, render } from '../../../../test';
+import { act, fireEvent, render } from 'testUtils';
 import App1 from '../demos/demo1';
 import App2 from '../demos/demo2';
