+++ packages/core/client/src/schema-component/antd/password/__tests__/password.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { fireEvent, render, screen } from '../../../../test';
+import { fireEvent, render, screen } from 'testUtils';
 import App1 from '../demos/demo1';
 import App2 from '../demos/demo2';
