+++ packages/core/client/src/schema-component/antd/percent/__tests__/percent.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { fireEvent, render, screen } from '../../../../test';
+import { fireEvent, render, screen } from 'testUtils';
 import App from '../demos/percent';
 
 describe('Percent', () => {
