--- packages/core/client/src/test/index.ts
@@ -1,3 +1,0 @@
-export * from './mockAPIClient';
-export * from './utils';
-
